+++ src/widget/text_widget.rs
@@ -15,9 +15,9 @@
 
 //extern crate graphics;
 
+use graphics;
 use graphics::character::CharacterCache;
 use graphics::draw_state::DrawState;
-use graphics;
 use graphics::*;
 use opengl_graphics::{GlGraphics, GlyphCache, TextureSettings};